+++ bundles/framework/admin-layerrights/Tile.js
@@ -47,7 +47,20 @@ Oskari.clazz.define('Oskari.framework.bundle.admin-layerrights.Tile',
          */
         startPlugin : function () {
             "use strict";
+            this._addTileStyleClasses();
             this.refresh();
+        },
+        _addTileStyleClasses: function() {
+            var isContainer = (this.container && this.instance.mediator) ? true : false;
+            var isBundleId = (isContainer && this.instance.mediator.bundleId) ? true : false;
+            var isInstanceId = (isContainer && this.instance.mediator.instanceId) ? true : false;
+
+            if (isInstanceId && !this.container.hasClass(this.instance.mediator.instanceId)) {
+                this.container.addClass(this.instance.mediator.instanceId);
+            }
+            if (isBundleId && !this.container.hasClass(this.instance.mediator.bundleId)) {
+                this.container.addClass(this.instance.mediator.bundleId);
+            }
         },
         /**
          * @method stopPlugin
@@ -95,17 +108,6 @@ Oskari.clazz.define('Oskari.framework.bundle.admin-layerrights.Tile',
          */
         refresh : function () {
             "use strict";
-            /*
-            var me = this,
-                instance = me.instance,
-                tpl = this.template,
-                sandbox = instance.getSandbox(),
-                cel = me.container,
-                status = cel.children('.oskari-tile-status');
-            */
-
-            // status.empty();
-            // status.append('(' + layers.length + ')');
         }
     }, {
         /**