+++ bundles/framework/divmanazer/instance.js
@@ -408,25 +408,11 @@ Oskari.clazz.define('Oskari.userinterface.bundle.ui.UserInterfaceBundleInstance'
                 } : null,
                 scroll: false,
                 stack: '.oskari-flyout',
-                create: function (event, ui) {
-                    // FIXME don't use jQuery.browser
-                    /* IE9 needs fixed width to not jump flyout width during and after dragging */
-                    if (jQuery.browser && jQuery.browser.msie && jQuery.browser.version[0] === '9') {
-                        flyout.css('width', flyout.width() + 'px');
-                    }
-                },
+                create: function (event, ui) {},
 
                 start: function () {
                     if (useHelper) {
                         flyout.css('display', 'none');
-                    } else {
-                        /* Attempt to fix IE9 vs. draggable flyout width issues */
-                        /* this did not work */
-                        /* if(jQuery.browser.msie && jQuery.browser.version[0] === "9") {
-                        flyout.css('width',flyout.width()+"px");
-                       }
-                       */
-
                     }
                 },