+++ bundles/framework/search/Tile.js
@@ -44,7 +44,20 @@ Oskari.clazz.define(
          * Interface method implementation, calls #refresh()
          */
         startPlugin: function () {
+            this._addTileStyleClasses();
             this.refresh();
+        },
+        _addTileStyleClasses: function() {
+            var isContainer = (this.container && this.instance.mediator) ? true : false;
+            var isBundleId = (isContainer && this.instance.mediator.bundleId) ? true : false;
+            var isInstanceId = (isContainer && this.instance.mediator.instanceId) ? true : false;
+
+            if (isInstanceId && !this.container.hasClass(this.instance.mediator.instanceId)) {
+                this.container.addClass(this.instance.mediator.instanceId);
+            }
+            if (isBundleId && !this.container.hasClass(this.instance.mediator.bundleId)) {
+                this.container.addClass(this.instance.mediator.bundleId);
+            }
         },
         /**
          * @method stopPlugin
@@ -86,18 +99,6 @@ Oskari.clazz.define(
          * Creates the UI for a fresh start
          */
         refresh: function () {
-            var me = this,
-                instance = me.instance,
-                cel = me.container,
-                idEl = cel.children('.oskari-tile-title'),
-                tpl = me.template,
-                sandbox = instance.getSandbox();
-
-            // var status = cel.children('.oskari-tile-status');
-            idEl.attr('id', 'oskari_search_tile_title');
-            //        status.empty();
-            //        status.append('(' + layers.length + ')');
-
         }
     }, {
         /**