+++ bundles/mapping/mapwfs2/service/Connection.js
@@ -117,22 +117,11 @@ Oskari.clazz.define(
          * Get browser information
          */
         getBrowser: function () {
-            // FIXME get rid of jQuery.browser,
-            // use some plugin if you must
+            var browser = Oskari.util.getBrowser();
             this.browser = {
                 name: 'unknown',
-                versionNum: parseInt(jQuery.browser.version, 10)
+                versionNum: parseInt(browser.version, 10)
             };
-
-            if (jQuery.browser.msie) {
-                this.browser.name = 'msie';
-            } else if (jQuery.browser.chrome) {
-                this.browser.name = 'chrome';
-            } else if (jQuery.browser.mozilla) {
-                this.browser.name = 'mozilla';
-            } else if (jQuery.browser.safari) {
-                this.browser.name = 'safari';
-            }
         },
 
         updateLazyDisconnect: function (isWFSOpen) {